requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dune_client==1.7.9
fonttools==4.58.4
frozenlist==1.7.0
idna==3.10
kaleido==1.0.0
kiwisolver==1.4.8
logistro==1.1.0
marshmallow==3.26.1
matplotlib==3.10.3
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.43.1
ndjson==0.3.1
numpy==2.3.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
propcache==0.3.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
seaborn==0.13.2
simplejson==3.20.1
six==1.17.0
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
types-setuptools==80.9.0.20250529
typing-inspect==0.9.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wrapt==1.17.2
yarl==1.20.1
